Rework of tool windows - LayerToolWindow
