+++ packages/main/src/react-app/LeftSidebar.tsx
@@ -1,26 +1,12 @@
 import { SideBar, SideBarPage } from './SideBar';
-import {
-  TbDatabaseSearch,
-  TbFile,
-  TbHistory,
-  TbPentagonPlus,
-  TbSearch,
-  TbStack,
-  TbTag
-} from 'react-icons/tb';
+import { TbDatabaseSearch, TbHistory, TbPentagonPlus, TbSearch, TbStack } from 'react-icons/tb';
 import { PickerToolWindow } from './toolwindow/PickerToolWindow/PickerToolWindow';
 import { LayerToolWindow } from './toolwindow/LayerToolWindow/LayerToolWindow';
-import { DocumentToolWindow } from './toolwindow/DocumentToolWindow/DocumentToolWindow';
 import { HistoryToolWindow } from './toolwindow/HistoryToolWindow/HistoryToolWindow';
 import { QueryToolWindow } from './toolwindow/QueryToolWindow/QueryToolWindow';
-import { makeActionMap } from '@diagram-craft/canvas/keyMap';
-import { defaultAppActions } from './appActionMap';
-import { useApplication } from '../application';
 import { DataToolWindow } from './toolwindow/DataToolWindow/DataToolWindow';
-import { TagsToolWindow } from './toolwindow/TagsToolWindow/TagsToolWindow';
 
 export const LeftSidebar = () => {
-  const application = useApplication();
   return (
     <SideBar side={'left'}>
       <SideBarPage icon={TbPentagonPlus}>
@@ -28,16 +14,6 @@ export const LeftSidebar = () => {
       </SideBarPage>
       <SideBarPage icon={TbStack}>
         <LayerToolWindow />
-      </SideBarPage>
-      <SideBarPage icon={TbFile}>
-        <DocumentToolWindow
-          document={application.model.activeDocument}
-          value={application.model.activeDiagram.id}
-          onValueChange={v => {
-            application.model.activeDiagram = application.model.activeDocument.byId(v)!;
-            application.actions = makeActionMap(defaultAppActions)(application);
-          }}
-        />
       </SideBarPage>
       <SideBarPage icon={TbHistory}>
         <HistoryToolWindow />
@@ -47,9 +23,6 @@ export const LeftSidebar = () => {
       </SideBarPage>
       <SideBarPage icon={TbDatabaseSearch}>
         <DataToolWindow />
-      </SideBarPage>
-      <SideBarPage icon={TbTag}>
-        <TagsToolWindow />
       </SideBarPage>
     </SideBar>
   );
