+++ packages/main/src/react-app/toolwindow/ObjectToolWindow/CanvasPanel.tsx
@@ -181,5 +181,5 @@ export const CanvasPanel = (props: Props) => {
 };
 
 type Props = {
-  mode?: 'accordion' | 'panel';
+  mode?: 'accordion' | 'panel' | 'headless';
 };
